requirements.txt
spotipy
python-dotenv
google-genai